zcash_client_sqlite: Move `TestState` to `zcash_client_backend`
