+++ zcash_client_sqlite/src/wallet/scanning.rs
@@ -587,6 +587,7 @@ pub(crate) mod tests {
     use zcash_client_backend::data_api::{
         chain::{ChainState, CommitmentTreeRoot},
         scanning::{spanning_tree::testing::scan_range, ScanPriority},
+        testing::{AddressType, FakeCompactOutput, InitialChainState, TestBuilder, TestState},
         AccountBirthday, Ratio, WalletRead, WalletWrite, SAPLING_SHARD_HEIGHT,
     };
     use zcash_primitives::{
@@ -601,7 +602,7 @@ pub(crate) mod tests {
         testing::{
             db::{TestDb, TestDbFactory},
             pool::ShieldedPoolTester,
-            AddressType, BlockCache, FakeCompactOutput, InitialChainState, TestBuilder, TestState,
+            BlockCache,
         },
         wallet::{
             sapling::tests::SaplingPoolTester,
